+++ pkg/splunk/enterprise/licensemanager.go
@@ -27,7 +27,6 @@ import (
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -281,40 +280,4 @@ func getLicenseManagerCurrentImage(ctx context.Context, c splcommon.ControllerCl
 	}
 
 	return "", nil
-}
-
-// changeClusterManagerAnnotations updates the checkUpdateImage field of the CLuster Manager Annotations to trigger the reconcile loop
-// on update, and returns error if something is wrong.
-func changeClusterManagerAnnotations(ctx context.Context, client splcommon.ControllerClient, cr *enterpriseApi.LicenseManager) error {
-
-	namespacedName := types.NamespacedName{
-		Namespace: cr.GetNamespace(),
-		Name:      cr.Spec.ClusterManagerRef.Name,
-	}
-	clusterManagerInstance := &enterpriseApi.ClusterManager{}
-	err := client.Get(context.TODO(), namespacedName, clusterManagerInstance)
-	if err != nil && k8serrors.IsNotFound(err) {
-		return nil
-	}
-	annotations := clusterManagerInstance.GetAnnotations()
-	if annotations == nil {
-		annotations = map[string]string{}
-	}
-	if _, ok := annotations["checkUpdateImage"]; ok {
-		if annotations["checkUpdateImage"] == clusterManagerInstance.Spec.Image {
-			return nil
-		}
-	}
-
-	annotations["checkUpdateImage"] = clusterManagerInstance.Spec.Image
-
-	clusterManagerInstance.SetAnnotations(annotations)
-	err = client.Update(ctx, clusterManagerInstance)
-	if err != nil {
-		fmt.Println("Error in Change Annotation UPDATE", err)
-		return err
-	}
-
-	return nil
-
 }
